requirements.txt
# App
django==4.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
django-extensions==3.2.3
ipython==8.18.1
pyotp==2.9.0
drf-yasg==1.21.7
pillow==10.3.0
qrcode==7.4.2
django-constance
channels
channels-redis

# Infra
uvicorn[standard]
psycopg2-binary>=2.9.1
redis